Remove examples for the moment
