+++ docs/ERP_api.rst
@@ -6,6 +6,7 @@ API
 Here is a quick guide to the layout of the package:
 
  - `erp/functions`: Functions used in pipelines
+ - `erp/drivers`: Drivers for pipelines
  - `tests`: Unit and regression tests
  - `docs`: Complete documentation. Includes non-interactive output of examples.
  - `data`: Test data used
@@ -17,6 +18,7 @@ The API is specified in the erp directory.
    :maxdepth: 1
 
    erp/functions/index.rst
+   erp/drivers/index.rst
 
 * :ref:`genindex`
 * :ref:`modindex`
