+++ docs/erp/drivers/index.rst
@@ -1,0 +1,18 @@
+.. _erp_drivers:
+
+.. py:currentmodule:: erp.drivers
+
+.. toctree::
+   :maxdepth: 2
+
+drivers
+*******
+
+.. automodapi::    erp.drivers.erp_script
+   :no-inheritance-diagram:
+
+.. automodapi::    erp.drivers.driver_support
+   :no-inheritance-diagram:
+
+
+
