+++ erp/functions/pipelines.py
@@ -4,18 +4,16 @@
 
 __all__ = ["run_erp"]
 
-import logging
-from functools import partial
 from erp.functions.imaging_steps import *
 from rascil.workflows.rsexecute.execution_support import rsexecute
 
-def run_erp(erp_params, get_logger=None):
+def run_erp(erp_params, get_logger):
     """ Run the Imaging/Selfcalibration pipeline
 
     :param erp_params: Control parameters
     :return:
     """
-
+    
     initialize_pipeline(erp_params, get_logger)
     
     bvis_list = ingest(erp_params)
