--- erp/functions/support.py
@@ -1,15 +1,0 @@
-""" Support functions needed for RASCIL pipeline
-
-"""
-
-__all__ = ['get_logger']
-
-import json
-import logging
-import os
-import shutil
-import sys
-import time
-
-logger = logging.getLogger('logger')
-
