--- examples/notebooks/default_params.json
@@ -1,131 +1,0 @@
-{
-  "global": {
-    "imaging_context"        : "2d",
-    "project"                : "eMERLIN_3C277_1",
-    "cmap"                   : "rainbow",
-    "results_directory"      : "./",
-    "initialize_pipeline"    : 1,
-    "list_ms"                : 1,
-    "load_ms"                : 1,
-    "plot_vis"               : 1,
-    "flag"                   : 1,
-    "average_channels"       : 1,
-    "combine_spw"            : 1,
-    "get_advice"             : 1,
-    "convert_stokesI"        : 1,
-    "create_images"          : 1,
-    "weight"                 : 1,
-    "cip"                    : 1,
-    "ical"                   : 1,
-    "write_images"           : 1,
-    "write_gaintables"       : 1,
-    "apply_calibration"      : 1,
-    "write_ms"               : 1,
-    "finalize_pipeline"      : 1
-  },
-  "dask": {
-    "distributed"            : true,
-    "nworkers"               : 4,
-    "memory"                 : 4e9
-  },
-  "list_ms": {
-    "ms_path"                 : "../../data/3C277.1C.16channels.ms",
-    "verbose"                 : true
-  },
-  "load_ms": {
-    "ms_path"                 : "../../data/3C277.1C.16channels.ms",
-    "verbose"                 : false,
-    "data_column"             : "DATA",
-    "source"                  : "1252+5634",
-    "dds"                     : [0]
-  },
-  "flag": {
-    "verbose"                 : false,
-    "strategy"                : "eMERLIN_strategy.rfis"
-  },
-  "plot_vis": {
-    "verbose"                 : true
-  },
-  "average_channels": {
-    "verbose"                 : true,
-    "original_nchan"          : 16,
-    "nchan"                   : 1
-  },
-  "combine_spw": {
-    "verbose"                 : true
-  },
-  "convert_stokesI": {
-    "verbose"                 : true
-  },
-  "get_advice": {
-    "verbose"                  : false,
-    "guard_band_image"         : 0.01,
-    "delA"                     : 0.02,
-    "oversampling_synthesised_beam" : 3.0,
-    "wprojection_planes"       : 1
-  },
-  "weight": {
-    "verbose"                 : false,
-    "algorithm"               : "uniform",
-    "robustness"              : 0.0
-  },
-  "create_images":  {
-    "verbose"                 : false,
-    "npixel"                  : 256,
-    "cellsize"                : 9e-08
-  },
-  "cip": {
-    "verbose"                 : true,
-    "nmajor"                 : 3,
-    "niter"                  : 1000,
-    "algorithm"              : "mmclean",
-    "nmoment"                : 2,
-    "gain"                   : 0.1,
-    "scales"                 : [0, 3, 10],
-    "threshold"              : 0.01,
-    "fractional_threshold"   : 0.3,
-    "window_shape"           : "quarter",
-    "do_wstacking"           : false
-  },
-  "ical": {
-    "verbose"                 : false,
-    "nmajor"                 : 8,
-    "niter"                  : 1000,
-    "algorithm"              : "mmclean",
-    "nmoment"                : 2,
-    "gain"                   : 0.1,
-    "scales"                 : [0, 3, 10],
-    "threshold"              : 0.003,
-    "fractional_threshold"   : 0.3,
-    "window_shape"           : "quarter",
-    "do_wstacking"           : false,
-    "T_first_selfcal"        : 2,
-    "T_phase_only"           : true,
-    "T_timeslice"            : "auto",
-    "G_first_selfcal"        : 5,
-    "G_phase_only"           : false,
-    "G_timeslice"            : 1200.0,
-    "B_first_selfcal"        : 8,
-    "B_phase_only"           : false,
-    "B_timeslice"            : 1e5,
-    "global_solution"        : true,
-    "calibration_context"    : "TG",
-    "do_selfcal"             : true,
-    "tol"                    : 1e-8
-  },
-  "write_images": {
-    "write_moments"          : true,
-    "number_moments"         : 1,
-    "verbose"                : true
-  },
-  "write_gaintables": {
-    "verbose"                : false
-  },
-  "apply_calibration": {
-    "verbose"                : true
-  },
-  "write_ms": {
-    "verbose"                : false,
-    "msout"                  : "../../test_results/3C277.1C.avg_test.ms"
-  }
-}
